chore: remove `compile_bin_package`
