+++ tooling/nargo_cli/src/cli/codegen_verifier_cmd.rs
@@ -1,12 +1,11 @@
+use super::fs::{create_named_dir, write_to_file};
 use super::NargoConfig;
-use super::{
-    compile_cmd::compile_bin_package,
-    fs::{create_named_dir, write_to_file},
-};
 use crate::backends::Backend;
+use crate::cli::compile_cmd::report_errors;
 use crate::errors::CliError;
 
 use clap::Args;
+use nargo::ops::compile_program;
 use nargo::{insert_all_files_for_workspace_into_file_manager, parse_all};
 use nargo_toml::{get_package_manifest, resolve_workspace_from_toml, PackageSelection};
 use noirc_driver::{file_manager_with_stdlib, CompileOptions, NOIR_ARTIFACT_VERSION_STRING};
@@ -47,13 +46,22 @@ pub(crate) fn run(
     let parsed_files = parse_all(&workspace_file_manager);
 
     let expression_width = backend.get_backend_info()?;
-    for package in &workspace {
-        let program = compile_bin_package(
+    let binary_packages = workspace.into_iter().filter(|package| package.is_binary());
+    for package in binary_packages {
+        let compilation_result = compile_program(
             &workspace_file_manager,
             &parsed_files,
             package,
             &args.compile_options,
             expression_width,
+            None,
+        );
+
+        let program = report_errors(
+            compilation_result,
+            &workspace_file_manager,
+            args.compile_options.deny_warnings,
+            args.compile_options.silence_warnings,
         )?;
 
         let smart_contract_string = backend.eth_contract(&program.circuit)?;
