+++ tooling/nargo_cli/src/cli/dap_cmd.rs
@@ -2,6 +2,7 @@ use acvm::acir::native_types::WitnessMap;
 use backend_interface::Backend;
 use clap::Args;
 use nargo::constants::PROVER_INPUT_FILE;
+use nargo::ops::compile_program;
 use nargo::workspace::Workspace;
 use nargo::{insert_all_files_for_workspace_into_file_manager, parse_all};
 use nargo_toml::{get_package_manifest, resolve_workspace_from_toml, PackageSelection};
@@ -21,7 +22,7 @@ use dap::server::Server;
 use dap::types::Capabilities;
 use serde_json::Value;
 
-use super::compile_cmd::compile_bin_package;
+use super::compile_cmd::report_errors;
 use super::fs::inputs::read_inputs_from_file;
 use crate::errors::CliError;
 
@@ -72,12 +73,21 @@ fn load_and_compile_project(
     insert_all_files_for_workspace_into_file_manager(&workspace, &mut workspace_file_manager);
     let parsed_files = parse_all(&workspace_file_manager);
 
-    let compiled_program = compile_bin_package(
+    let compile_options = CompileOptions::default();
+    let compilation_result = compile_program(
         &workspace_file_manager,
         &parsed_files,
         package,
-        &CompileOptions::default(),
+        &compile_options,
         expression_width,
+        None,
+    );
+
+    let compiled_program = report_errors(
+        compilation_result,
+        &workspace_file_manager,
+        compile_options.deny_warnings,
+        compile_options.silence_warnings,
     )
     .map_err(|_| LoadError("Failed to compile project"))?;
