+++ tooling/nargo_cli/src/cli/debug_cmd.rs
@@ -6,6 +6,7 @@ use clap::Args;
 
 use nargo::artifacts::debug::DebugArtifact;
 use nargo::constants::PROVER_INPUT_FILE;
+use nargo::ops::compile_program;
 use nargo::package::Package;
 use nargo::{insert_all_files_for_workspace_into_file_manager, parse_all};
 use nargo_toml::{get_package_manifest, resolve_workspace_from_toml, PackageSelection};
@@ -16,7 +17,7 @@ use noirc_driver::{
 };
 use noirc_frontend::graph::CrateName;
 
-use super::compile_cmd::compile_bin_package;
+use super::compile_cmd::report_errors;
 use super::fs::{inputs::read_inputs_from_file, witness::save_witness_to_dir};
 use super::NargoConfig;
 use crate::backends::Backend;
@@ -66,12 +67,20 @@ pub(crate) fn run(
         return Ok(());
     };
 
-    let compiled_program = compile_bin_package(
+    let compilation_result = compile_program(
         &workspace_file_manager,
         &parsed_files,
         package,
         &args.compile_options,
         expression_width,
+        None,
+    );
+
+    let compiled_program = report_errors(
+        compilation_result,
+        &workspace_file_manager,
+        args.compile_options.deny_warnings,
+        args.compile_options.silence_warnings,
     )?;
 
     run_async(package, compiled_program, &args.prover_name, &args.witness_name, target_dir)
